requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.2
alpyvantage==0.2.1
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.6.1
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.3
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
aws-cfn-bootstrap==2.0
awscli==2.9.19
awscrt==0.16.7
Babel==2.12.1
blinker==1.6.2
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
chevron==0.13.1
click==8.1.6
clickclick==20.10.2
cloud-init==22.2.2
colorama==0.4.6
colorlog==4.8.0
configobj==5.0.6
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
dbus-python==1.2.18
Deprecated==1.2.14
dill==0.3.7
distro==1.5.0
dnspython==2.4.1
docutils==0.20.1
ec2-hibinit-agent==1.0.2
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.4.0
google-re2==1.1
gpg==1.15.1
graphviz==0.20.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
libcomps==0.1.18
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
netifaces==0.10.6
numpy==1.25.2
oauthlib==3.0.2
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.2.0
ply==3.11
prettytable==0.7.2
prison==0.2.1
prompt-toolkit==3.0.24
psutil==5.9.5
pyarrow==12.0.1
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
release-notification==1.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.5.2
rich-argparse==1.2.0
rpds-py==0.9.2
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
support-info==1.0
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==2.0.4
wcwidth==0.2.5
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
